+++ src/Controller/FormValidator/FormAuthenticationValidator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\FormValidator;
 
 use Core\DefaultAbstract\FormValidatorAbstract;
