+++ src/Controller/RegisterController.php
@@ -2,13 +2,13 @@
 
 namespace App\Controller;
 
+use App\Controller\FormValidator\FormRegisterValidator;
 use App\Entity\User;
 use App\Repository\UserRepository;
 use App\utils\Helper;
 use Core\DefaultAbstract\DefaultAbstractController;
 use Core\Exception\CoreException;
 use Core\Traits\Controller\AddControllerTrait;
-use Exception;
 
 /**
  * Class RegisterController
@@ -58,8 +58,9 @@ class RegisterController extends DefaultAbstractController
 
     /**
      * @param array|mixed[] $formValues
-     * @param object $entity
-     * @throws CoreException
+     * @param object        $entity
+     *
+     * @throws CoreException*@throws Exception
      */
     public function check(array $formValues, $entity): void
     {
@@ -67,7 +68,7 @@ class RegisterController extends DefaultAbstractController
         $password = $formValues['password'] ?? '';
 
         if (false === Helper::checkEmail($email)) {
-            throw new Exception("l'adresse $email n'est pas valide");
+            throw new CoreException("l'adresse $email n'est pas valide");
         }
 
         if ($formValues['password'] !== $formValues['password2']) {
